kafka/producer/requirements.txt
confluent-kafka==2.4.0
websocket-client==1.8.0
python-dotenv==1.0.1
